internment
Struct Intern

pub struct Intern<T: 'static + ?Sized> { /* private fields */ }

A pointer to an interned object.

The interned object will be held in memory indefinitely. On the plus side, this means that lifetime issues are simple when using Intern.

Example

use internment::Intern;

let x = Intern::new("hello");
let y = Intern::new("world");
assert_ne!(x, y);
assert_eq!(x, Intern::new("hello"));
assert_eq!(*x, "hello"); // dereference an Intern like a pointer

Example with owned String data

use internment::Intern;

let x = Intern::new("hello".to_string());
let y = Intern::<String>::from_ref("world");
assert_ne!(x, y);
assert_eq!(x, Intern::from_ref("hello"));
assert_eq!(y, Intern::from_ref("world"));
assert_eq!(&*x, "hello"); // dereference a Intern like a pointer

Implementations

impl<T: Eq + Hash + Send + Sync + 'static> Intern<T>

pub fn new(val: T) -> Intern<T>

Intern a value.

If this value has not previously been interned, then new will allocate a spot for the value on the heap. Otherwise, it will return a pointer to the object previously allocated.

Note that Intern::new is a bit slow, since it needs to check a HashSet protected by a Mutex.

pub fn from_ref<'a, Q: ?Sized + Eq + Hash + 'a>(val: &'a Q) -> Intern<T>
where T: Borrow<Q> + From<&'a Q>,

Intern a value from a reference.

If this value has not previously been interned, then new will allocate a spot for the value on the heap and generate that value using T::from(val).

impl<T: Eq + Hash + Send + Sync + 'static + ?Sized> Intern<T>

pub fn as_ref(self) -> &'static T

Get a long-lived reference to the data pointed to by an Intern, which is never freed from the intern pool.

pub fn num_objects_interned() -> usize

See how many objects have been interned. This may be helpful in analyzing memory use.

pub fn is_interned<'a, Q: ?Sized + Eq + Hash + 'a>(val: &'a Q) -> bool
where T: Borrow<Q>,

Check if a value already is interned.

If this value has previously been interned, return true, else returns false/// Checking if an object is already interned

 
use internment::Intern;

assert!(!Intern::<String>::is_interned("Fortunato"));
let x = Intern::new("Fortunato".to_string());
assert!(Intern::<String>::is_interned("Fortunato"));

assert!(!Intern::<str>::is_interned("Fortunato"));
let x: Intern<str> = "Fortunato".into();
assert!(Intern::<str>::is_interned("Fortunato"));